requirements.txt
APScheduler==3.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.51.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
fonttools==4.37.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.23.3
oauthlib==3.2.0
openapi-codec==1.3.2
outcome==1.2.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.4.3
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
wsproto==1.2.0
zipp==3.8.1
